requirements.txt
openai
anthropic
deepseek-sdk
transformers
accelerate
langchain
llama-index
pandas
numpy
streamlit
python-dotenv
tenacity
